+++ package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/dayekimm0/Rookie#readme",
   "dependencies": {
+    "@emotion/is-prop-valid": "^1.4.0",
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-regular-svg-icons": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
