+++ src/components/TeamHome/InfSlide.jsx
@@ -1,4 +1,4 @@
-import { useState, useMemo, useCallback, useEffect } from "react";
+import { useState, useMemo, useCallback, useEffect, memo } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import { Swiper, SwiperSlide } from "swiper/react";
@@ -10,7 +10,6 @@ import {
   useYoutubePlaylist,
   useYoutubeVideoDetails,
 } from "../../hook/useYoutubePlayList";
-import Spinner from "../Spinner";
 
 const Container = styled.div`
   position: relative;
@@ -48,24 +47,13 @@ const SlideContainer = styled.div`
   }
 `;
 
-const SpinnerWrap = styled.div`
-  padding: 80px 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const InfSlide = ({ playlistId, max }) => {
+const InfSlide = memo(({ playlistId, max }) => {
   const [swiper, setSwiper] = useState();
   const [isBeginning, setIsBeginning] = useState(true);
   const [isEnd, setIsEnd] = useState(false);
   const navigate = useNavigate();
 
-  const {
-    data: plays = [],
-    isLoading,
-    isError,
-  } = useYoutubePlaylist(playlistId, max);
+  const { data: plays = [] } = useYoutubePlaylist(playlistId, max, true, true);
 
   const videoIds = useMemo(() => {
     return plays
@@ -78,7 +66,11 @@ const InfSlide = ({ playlistId, max }) => {
     navigate(`/play/${videoId}`);
   };
 
-  const { data: details = [] } = useYoutubeVideoDetails(videoIds, !!videoIds);
+  const { data: details = [] } = useYoutubeVideoDetails(
+    videoIds,
+    !!videoIds,
+    true
+  );
 
   const slides = useMemo(() => {
     return details.map((video) => {
@@ -115,14 +107,6 @@ const InfSlide = ({ playlistId, max }) => {
       setIsEnd(swiper.isEnd);
     }
   }, [swiper, details]);
-
-  if (isLoading)
-    return (
-      <SpinnerWrap>
-        <Spinner />
-      </SpinnerWrap>
-    );
-  if (isError) return <div>문제가 발생하였습니다.</div>;
 
   return (
     <Container>
@@ -186,6 +170,6 @@ const InfSlide = ({ playlistId, max }) => {
       </SlideContainer>
     </Container>
   );
-};
+});
 
 export default InfSlide;
