+++ src/components/TeamHome/UpcomingMatch.jsx
@@ -1,3 +1,4 @@
+import { useMemo } from "react";
 import styled from "styled-components";
 import scheduleData from "../../data/gameList_final.json";
 import MatchCard from "./MatchCard";
@@ -42,47 +43,47 @@ const Wrapper = styled.div`
 `;
 
 const UpcomingMatch = ({ teamCode }) => {
-  const today = new Date().toISOString().split("T")[0];
-  const baseIndex = scheduleData.findIndex((item) => item.date >= today);
-  const safeIndex = baseIndex !== -1 ? baseIndex : scheduleData.length - 1;
+  const threeDaySlice = useMemo(() => {
+    const today = new Date().toISOString().split("T")[0];
+    const baseIndex = scheduleData.findIndex((item) => item.date >= today);
+    const safeIndex = baseIndex !== -1 ? baseIndex : scheduleData.length - 1;
 
-  // 첫째날, 마지막날 일정 체크
-  const isFirstDay = safeIndex === 0;
-  const isLastDay = safeIndex === scheduleData.length - 1;
+    // 첫째날, 마지막날 일정 체크
+    const isFirstDay = safeIndex === 0;
+    const isLastDay = safeIndex === scheduleData.length - 1;
 
-  // const threeDaySlice = [
-  //   scheduleData[safeIndex],
-  //   scheduleData[safeIndex + 1],
-  //   scheduleData[safeIndex + 2],
-  // ].filter(Boolean);
+    // 보여줄 3일 계산
+    const displayedDays = isLastDay
+      ? [
+          scheduleData[safeIndex - 2],
+          scheduleData[safeIndex - 1],
+          scheduleData[safeIndex],
+        ]
+      : isFirstDay
+      ? [
+          scheduleData[safeIndex],
+          scheduleData[safeIndex + 1],
+          scheduleData[safeIndex + 2],
+        ]
+      : [
+          scheduleData[safeIndex - 1],
+          scheduleData[safeIndex],
+          scheduleData[safeIndex + 1],
+        ];
 
-  // 보여줄 3일 계산
-  const displayedDays = isLastDay
-    ? [
-        scheduleData[safeIndex - 2],
-        scheduleData[safeIndex - 1],
-        scheduleData[safeIndex],
-      ]
-    : isFirstDay
-    ? [
-        scheduleData[safeIndex],
-        scheduleData[safeIndex + 1],
-        scheduleData[safeIndex + 2],
-      ]
-    : [
-        scheduleData[safeIndex - 1],
-        scheduleData[safeIndex],
-        scheduleData[safeIndex + 1],
-      ];
-
-  const threeDaySlice = displayedDays.filter(Boolean);
+    return displayedDays.filter(Boolean);
+  }, []);
 
   return (
     <Container className="inner">
       <h3 className="title">UPCOMING MATCH</h3>
       <Wrapper>
         {threeDaySlice.map((match, index) => (
-          <MatchCard key={`match${index}`} match={match} teamCode={teamCode} />
+          <MatchCard
+            key={match.date || `match-${index}`}
+            match={match}
+            teamCode={teamCode}
+          />
         ))}
       </Wrapper>
     </Container>
